Move new tests to separated files.
